restructuring reth-node-metrics crate
