+++ crates/cli/commands/src/node.rs
@@ -15,6 +15,7 @@ use reth_node_core::{
     node_config::NodeConfig,
     version,
 };
+use reth_node_metrics::recorder::PROMETHEUS_RECORDER_HANDLE;
 use std::{ffi::OsString, fmt, future::Future, net::SocketAddr, path::PathBuf, sync::Arc};
 
 /// Start the node
@@ -173,7 +174,7 @@ impl<Ext: clap::Args + fmt::Debug> NodeCommand<Ext> {
 
         // Register the prometheus recorder before creating the database,
         // because database init needs it to register metrics.
-        let _ = node_config.install_prometheus_recorder()?;
+        let _ = PROMETHEUS_RECORDER_HANDLE.clone();
 
         let data_dir = node_config.datadir();
         let db_path = data_dir.db();
