+++ crates/ethereum/node/Cargo.toml
@@ -34,7 +34,6 @@ reth-tasks.workspace = true
 reth-tokio-util.workspace = true
 reth-node-events.workspace = true
 reth-node-core.workspace = true
-reth-node-metrics.workspace = true
 reth-exex.workspace = true
 reth-blockchain-tree.workspace = true
