+++ crates/node/metrics/Cargo.toml
@@ -27,7 +27,6 @@ http.workspace = true
 tower.workspace = true
 
 tracing.workspace = true
-const_format.workspace = true
 eyre.workspace = true
 
 [target.'cfg(unix)'.dependencies]
